pyproject.toml
[build-system]
requires = ['setuptools']
build-backend = 'setuptools.build_meta'

[project]
name = "mimebytes"
version = "0.2.0"
description = "Make repr of bytes better in IPython"
readme = "README.md"
authors = [
    {name = "Bane Sullivan", email = "hello@banesullivan.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = '>=3.8'

[project.optional-dependencies]
magic = [
    "python-magic",
]

[project.urls]
"Bug Tracker" = 'https://github.com/banesullivan/mimebytes/issues'
"Source Code" = 'https://github.com/banesullivan/mimebytes'

[tool.black]
line-length = 100
skip-string-normalization = false
target-version = ["py38"]
exclude='\.eggs|\.git|\.mypy_cache|\.tox|\.venv|_build|buck-out|build|dist|node_modules'

[tool.isort]
profile = "black"
line_length = 100
# Sort by name, don't cluster "from" vs "import"
force_sort_within_sections = true
# Combines "as" imports on the same line
combine_as_imports = true
